Refactor to make things easier to understand
